refactor: extract project_media.py and clip_scanner.py from backend
Split backend/project.py (743 lines):
- project_media.py: sequence/media project creation functions
- project.py: core project management, JSON helpers, path utilities

Split backend/clip_state.py (670 lines):
- clip_scanner.py: scan_clips_dir, scan_project_clips
- clip_state.py: ClipEntry, ClipState, state machine, dataclasses

Backward-compat re-exports in original modules. All 427 tests pass.
