+++ backend/__init__.py
@@ -3,8 +3,8 @@
 from .clip_state import (
     ClipAsset, ClipEntry, ClipState, InOutRange, PipelineRoute,
     classify_pipeline_route, mask_sequence_is_videomama_ready,
-    scan_clips_dir, scan_project_clips,
 )
+from .clip_scanner import scan_clips_dir, scan_project_clips
 from .errors import CorridorKeyError
 from .job_queue import GPUJob, GPUJobQueue, JobType, JobStatus
 from .project import (
@@ -16,6 +16,8 @@ from .project import (
     is_video_file, is_image_file, VIDEO_FILE_FILTER,
     folder_has_image_sequence, count_sequence_frames,
     validate_sequence_stems, find_clip_by_source,
+)
+from .project_media import (
     create_clip_from_sequence,
     add_sequences_to_project, create_project_from_media,
 )
